added ShoppingCart entity and all necessary layers fot it

+++ src/main/java/com/products/test/model/User.java
@@ -1,16 +1,15 @@
 package com.products.test.model;
 
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.Table;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
 @Data
 @Builder
@@ -25,5 +24,5 @@ public class User {
     private Long id;
     private String email;
     private String password;
-    private int moneyAccount;
+    private double moneyAccount;
 }
